users_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.8
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
